CarbMine/backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
joblib==1.4.2
firebase-admin==6.5.0
python-multipart==0.0.9

